--- src/main/java/org/lucky0111/pettalk/controller/MainController.java
@@ -1,16 +1,0 @@
-package org.lucky0111.pettalk.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ResponseBody;
-
-@Controller
-public class MainController {
-
-    @GetMapping("/")
-    @ResponseBody
-    public String mainAPI() {
-
-        return "main route";
-    }
-}
