+++ src/main/java/org/lucky0111/pettalk/domain/dto/error/ErrorResponseDTO.java
@@ -1,7 +1,6 @@
-package org.lucky0111.pettalk.domain.dto;
+package org.lucky0111.pettalk.domain.dto.error;
 
 import lombok.Builder;
-import lombok.Getter;
 
 @Builder
 public record ErrorResponseDTO (
