+++ src/main/java/org/lucky0111/pettalk/domain/entity/auth/RefreshToken.java
@@ -1,10 +1,11 @@
-package org.lucky0111.pettalk.domain.entity;
+package org.lucky0111.pettalk.domain.entity.auth;
 
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import org.lucky0111.pettalk.domain.common.BaseTimeEntity;
+import org.lucky0111.pettalk.domain.entity.user.PetUser;
 
 import java.time.LocalDateTime;
