+++ src/main/java/org/lucky0111/pettalk/domain/entity/common/PetCategory.java
@@ -1,4 +1,4 @@
-package org.lucky0111.pettalk.domain.entity;
+package org.lucky0111.pettalk.domain.entity.common;
 
 import jakarta.persistence.*;
 import lombok.Getter;
