+++ src/main/java/org/lucky0111/pettalk/domain/entity/review/ReviewLike.java
@@ -3,7 +3,7 @@ package org.lucky0111.pettalk.domain.entity.review;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
-import org.lucky0111.pettalk.domain.entity.PetUser;
+import org.lucky0111.pettalk.domain.entity.user.PetUser;
 
 @Setter
 @Getter
