+++ src/main/java/org/lucky0111/pettalk/domain/entity/trainer/Certification.java
@@ -1,4 +1,4 @@
-package org.lucky0111.pettalk.domain.entity;
+package org.lucky0111.pettalk.domain.entity.trainer;
 
 import jakarta.persistence.*;
 import lombok.Getter;
