+++ src/main/java/org/lucky0111/pettalk/domain/entity/trainer/TrainerTagRelation.java
@@ -1,7 +1,8 @@
-package org.lucky0111.pettalk.domain.entity;
+package org.lucky0111.pettalk.domain.entity.trainer;
 
 import jakarta.persistence.*;
 import lombok.Getter;
+import org.lucky0111.pettalk.domain.entity.common.Tag;
 
 @Getter
 @Entity
