+++ src/main/java/org/lucky0111/pettalk/domain/entity/user/PetUser.java
@@ -1,4 +1,4 @@
-package org.lucky0111.pettalk.domain.entity;
+package org.lucky0111.pettalk.domain.entity.user;
 
 import jakarta.persistence.*;
 import lombok.Getter;
