+++ src/main/java/org/lucky0111/pettalk/repository/common/TagRepository.java
@@ -1,6 +1,6 @@
 package org.lucky0111.pettalk.repository.common;
 
-import org.lucky0111.pettalk.domain.entity.Tag;
+import org.lucky0111.pettalk.domain.entity.common.Tag;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
