+++ src/main/java/org/lucky0111/pettalk/repository/trainer/CertificationRepository.java
@@ -1,8 +1,7 @@
 package org.lucky0111.pettalk.repository.trainer;
 
-import org.lucky0111.pettalk.domain.entity.Certification;
+import org.lucky0111.pettalk.domain.entity.trainer.Certification;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 
 import java.util.List;
 import java.util.UUID;
