+++ src/main/java/org/lucky0111/pettalk/repository/trainer/TrainerRepository.java
@@ -1,9 +1,8 @@
 package org.lucky0111.pettalk.repository.trainer;
 
-import org.lucky0111.pettalk.domain.entity.PetUser;
-import org.lucky0111.pettalk.domain.entity.Trainer;
+import org.lucky0111.pettalk.domain.entity.user.PetUser;
+import org.lucky0111.pettalk.domain.entity.trainer.Trainer;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 import java.util.UUID;
