+++ src/main/java/org/lucky0111/pettalk/service/auth/CustomOAuth2UserService.java
@@ -7,7 +7,7 @@ import org.lucky0111.pettalk.domain.dto.auth.KakaoResponse;
 import org.lucky0111.pettalk.domain.dto.auth.NaverResponse;
 import org.lucky0111.pettalk.domain.dto.auth.OAuth2Response;
 import org.lucky0111.pettalk.domain.dto.user.UserDTO;
-import org.lucky0111.pettalk.domain.entity.PetUser;
+import org.lucky0111.pettalk.domain.entity.user.PetUser;
 import org.lucky0111.pettalk.repository.user.PetUserRepository;
 import org.springframework.security.oauth2.client.userinfo.DefaultOAuth2UserService;
 import org.springframework.security.oauth2.client.userinfo.OAuth2UserRequest;
