+++ src/main/java/org/lucky0111/pettalk/service/match/UserApplyServiceImpl.java
@@ -6,8 +6,8 @@ import lombok.extern.slf4j.Slf4j;
 import org.lucky0111.pettalk.domain.common.Status;
 import org.lucky0111.pettalk.domain.dto.match.UserApplyRequestDTO;
 import org.lucky0111.pettalk.domain.dto.match.UserApplyResponseDTO;
-import org.lucky0111.pettalk.domain.entity.PetUser;
-import org.lucky0111.pettalk.domain.entity.Trainer;
+import org.lucky0111.pettalk.domain.entity.user.PetUser;
+import org.lucky0111.pettalk.domain.entity.trainer.Trainer;
 import org.lucky0111.pettalk.domain.entity.match.UserApply;
 import org.lucky0111.pettalk.repository.match.UserApplyRepository;
 import org.lucky0111.pettalk.repository.trainer.TrainerRepository;
