+++ src/main/java/org/lucky0111/pettalk/service/trainer/TrainerServiceImpl.java
@@ -3,7 +3,11 @@ package org.lucky0111.pettalk.service.trainer;
 import lombok.RequiredArgsConstructor;
 import org.lucky0111.pettalk.domain.dto.trainer.CertificationDTO;
 import org.lucky0111.pettalk.domain.dto.trainer.TrainerDTO;
-import org.lucky0111.pettalk.domain.entity.*;
+import org.lucky0111.pettalk.domain.entity.user.PetUser;
+import org.lucky0111.pettalk.domain.entity.common.Tag;
+import org.lucky0111.pettalk.domain.entity.trainer.Certification;
+import org.lucky0111.pettalk.domain.entity.trainer.Trainer;
+import org.lucky0111.pettalk.domain.entity.trainer.TrainerTagRelation;
 import org.lucky0111.pettalk.exception.CustomException;
 import org.lucky0111.pettalk.repository.common.TagRepository;
 import org.lucky0111.pettalk.repository.review.ReviewRepository;
