+++ src/main/java/org/lucky0111/pettalk/util/auth/JWTUtil.java
@@ -5,8 +5,8 @@ import io.jsonwebtoken.Jwts;
 import jakarta.transaction.Transactional;
 import org.lucky0111.pettalk.domain.dto.auth.OAuthTempTokenDTO;
 import org.lucky0111.pettalk.domain.dto.auth.TokenDTO;
-import org.lucky0111.pettalk.domain.entity.PetUser;
-import org.lucky0111.pettalk.domain.entity.RefreshToken;
+import org.lucky0111.pettalk.domain.entity.user.PetUser;
+import org.lucky0111.pettalk.domain.entity.auth.RefreshToken;
 import org.lucky0111.pettalk.repository.auth.RefreshTokenRepository;
 import org.lucky0111.pettalk.repository.user.PetUserRepository;
 import org.springframework.beans.factory.annotation.Value;
